+++ src/uucp__break.ml
@@ -28,6 +28,12 @@ module Low = struct
 
   let sentence_max = sentence_max
   let sentence_of_int = sentence_of_byte
+
+  let indic_conjunct_break u =
+    Uucp_tmapbyte.get Uucp_break_data.indic_conjunct_break_map (Uchar.to_int u)
+
+  let indic_conjunct_break_max = indic_conjunct_break_max
+  let indic_conjunct_break_of_int = indic_conjunct_break_of_byte
 end
 
 let line u = Array.unsafe_get Low.line_of_int (Low.line u)
@@ -36,6 +42,9 @@ let grapheme_cluster u = Array.unsafe_get Low.grapheme_cluster_of_int
 
 let word u = Array.unsafe_get Low.word_of_int (Low.word u)
 let sentence u = Array.unsafe_get Low.sentence_of_int (Low.sentence u)
+let indic_conjunct_break u =
+  Array.unsafe_get Low.indic_conjunct_break_of_int
+    (Low.indic_conjunct_break u)
 
 let east_asian_width u =
   Uucp_rmap.get Uucp_break_data.east_asian_width_map (Uchar.to_int u)