+++ test/perf.ml
@@ -38,6 +38,7 @@ let lookup_break count =
   prop "grapheme_cluster" Uucp.Break.grapheme_cluster;
   prop "word" Uucp.Break.word;
   prop "sentence" Uucp.Break.sentence;
+  prop "indic_conjunct_break" Uucp.Break.indic_conjunct_break;
   prop "east_asian_width" Uucp.Break.east_asian_width;
   ()